backend/services/api-gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
redis==5.0.1
aioredis==2.0.1
asyncpg==0.29.0
httpx==0.25.2
python-dotenv==1.0.0
prometheus-client==0.19.0
slowapi==0.1.9
structlog==23.2.0